Added in external sleep pin, cleaned up some code in Main.c. Changed analogue API and swapped over uint8_t string routines and buffers to char to allow for a generic makefile.

+++ V2Protocol.c
@@ -12,7 +12,7 @@ const uint8_t SignonResponse[]  PROGMEM = {AICB_CMD_SIGN_ON, AICB_STATUS_CMD_OK,
 
 // GLOBAL VARIABLES:
 uint8_t  PacketBytes[V2P_MAXBUFFSIZE]   = {};
-uint16_t SequenceNum                    = 0;
+uint8_t  SequenceNum                    = 0;
 uint16_t MessageSize                    = 0;
 uint8_t  InProgrammingMode              = FALSE;
 uint32_t CurrAddress                    = 0;
